requirements.txt
requests>=2.31.0
psycopg2-binary>=2.9.9
qdrant-client>=1.7.0
python-dotenv>=1.0.0
fastapi>=0.109.0
uvicorn>=0.27.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.1.0
email-validator>=2.1.0
python-multipart>=0.0.6
